Created private report generation function
